move AsyncReducerStatus where it belongs
