+++ applications/fishing-map/src/features/datasets/NewDatasetTooltip.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { batch, useDispatch, useSelector } from 'react-redux'
 import Spinner from '@globalfishingwatch/ui-components/dist/spinner'
 import { Dataset, DatasetCategory } from '@globalfishingwatch/api-types'
-import { AsyncReducerStatus } from 'types'
+import { AsyncReducerStatus } from 'utils/async-slice'
 import { selectUserDatasetsNotUsed } from 'features/user/user.selectors'
 import { useDatasetModalConnect, useNewDatasetConnect } from './datasets.hook'
 import styles from './NewDatasetTooltip.module.css'
