+++ applications/fishing-map/src/features/sidebar/Sidebar.tsx
@@ -14,7 +14,7 @@ import User from 'features/user/User'
 import Workspace from 'features/workspace/Workspace'
 import WorkspacesList from 'features/workspaces-list/WorkspacesList'
 import NewDataset from 'features/datasets/NewDataset'
-import { AsyncReducerStatus } from 'types'
+import { AsyncReducerStatus } from 'utils/async-slice'
 import { selectHighlightedWorkspacesStatus } from 'features/workspaces-list/workspaces-list.slice'
 import styles from './Sidebar.module.css'
 import CategoryTabs from './CategoryTabs'
