+++ applications/fishing-map/src/features/workspace/vessels/VesselLayerPanel.tsx
@@ -15,7 +15,8 @@ import {
   filterSegmentsByTimerange,
 } from '@globalfishingwatch/data-transforms'
 import { formatInfoField } from 'utils/info'
-import { UrlDataviewInstance, AsyncReducerStatus } from 'types'
+import { UrlDataviewInstance } from 'types'
+import { AsyncReducerStatus } from 'utils/async-slice'
 import styles from 'features/workspace/shared/LayerPanel.module.css'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { resolveDataviewDatasetResource } from 'features/workspace/workspace.selectors'
