+++ applications/fishing-map/src/types/index.ts
@@ -49,18 +49,6 @@ export type MapCoordinates = {
   transitionDuration?: number
 }
 
-export enum AsyncReducerStatus {
-  Idle = 'idle',
-  Aborted = 'aborted',
-  Loading = 'loading',
-  LoadingItem = 'loading.item',
-  LoadingCreate = 'loading.create',
-  LoadingUpdate = 'loading.update',
-  LoadingDelete = 'loading.delete',
-  Finished = 'finished',
-  Error = 'error',
-}
-
 export enum TimebarVisualisations {
   Heatmap = 'heatmap',
   Vessel = 'vessel',
